refactor: RandomNumberGenerator -> RandomPowerGenerator 도메인으로 리팩토링
