+++ src/main/java/domain/RandomPowerGenerator.java
@@ -1,10 +1,11 @@
-package utils;
+package domain;
 
 import java.util.Random;
 
-public class RandomNumberGenerator {
+public class RandomPowerGenerator {
     private static final int RANDOM_NUMBER_UPPER_BOUND = 10;
-    public static int generate() {
+
+    public int generate() {
         Random random = new Random();
         return random.nextInt(RANDOM_NUMBER_UPPER_BOUND);
     }
